pyproject.toml
[project]
name = "free-parc-app"
version = "0.1.0"
description = ""
authors = [
    {name = "DaiTensa",email = "dai.tensaout@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.44.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (<4.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
